[core] Fix potential precision data loss

+++ src/mbgl/source/source_info.hpp
@@ -19,8 +19,8 @@ public:
     uint16_t minZoom = 0;
     uint16_t maxZoom = 22;
     std::string attribution;
-    std::array<float, 3> center = { { 0, 0, 0 } };
-    std::array<float, 4> bounds = { { -180, -90, 180, 90 } };
+    std::array<double, 3> center = { { 0, 0, 0 } };
+    std::array<double, 4> bounds = { { -180, -90, 180, 90 } };
 };
 
 } // namespace mbgl
